+++ Assets/KimChang/KimScript/Test/TestModeUI.cs
@@ -145,7 +145,7 @@ public class TestModeUI : MonoBehaviour
         {
             if (int.TryParse(match.Value, out int number))
             {
-                if (number >= 0 && number <= 25)
+                if (number >= 0 && number <= 66)
                 {
                     enemyUnitIds.Add(number);
                 }
